requirements.txt
tensorflow>=2.13.0
numpy>=1.24.3
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.2
seaborn>=0.12.2
gradio>=4.1.1
albumentations>=1.3.1